adding dashboard and updating validation
