+++ src/App.js
@@ -1,19 +1,10 @@
-import { Box } from "@chakra-ui/react";
-import UserInputModal from "./components/UserInputModal";
-import { useState } from "react";
+import { Dashboard } from "./components/Dashboard";
 
 function App() {
-  const [initialObject, setInitialObject] = useState(null);
-  const [patches, setPatches] = useState([]);
-
-  const handleSubmit = (object, patches) => {
-    setInitialObject(object);
-    setPatches(patches);
-  };
   return (
-    <Box>
-      <UserInputModal onSubmit={handleSubmit} />
-    </Box>
+    <>
+      <Dashboard />
+    </>
   );
 }
