+++ src/utils/validation.js
@@ -1,15 +1,6 @@
-export const isValidJsonObject = (str) => {
+export const isValidJsObject = (input) => {
   try {
-      const parsed = JSON.parse(str);
-
-      console.log('str0000',parsed);
-      console.log('typeof parsed === "object"',typeof parsed === "object")
-      console.log("parsed !== null",parsed !== null)
-      console.log('!Array.isArray(parsed)',!Array.isArray(parsed))
-      console.log('Object.keys(parsed).length > 0',Object.keys(parsed).length > 0)
-      console.log('false',str);
-  
-
+    const parsed = convertStringToJsObject(input);
     return (
       typeof parsed === "object" &&
       parsed !== null &&
@@ -17,11 +8,16 @@ export const isValidJsonObject = (str) => {
       Object.keys(parsed).length > 0
     );
   } catch (error) {
-    console.log('typeof parsed === "object"---',typeof str === "object")
-    console.log("parsed !== null",str !== null)
-    console.log('!Array.isArray(parsed)',!Array.isArray(str))
-    console.log('Object.keys(parsed).length > 0',Object.keys(str).length > 0)
-    console.log('false',str);
     return false;
+  }
+};
+
+export const convertStringToJsObject = (str) => {
+  try {
+    const parsed = new Function(`return ${str};`)();
+    return parsed;
+  } catch (error) {
+    console.error("Error parsing string to JS object:", error);
+    return null;
   }
 };
